+++ server/src/main/java/com/blocklang/core/service/DocumentService.java
@@ -1,0 +1,9 @@
+package com.blocklang.core.service;
+
+import java.util.Optional;
+
+public interface DocumentService {
+
+	Optional<String> findByFileName(String fileName);
+
+}